backend/requirements.txt
django>=3.2.4,<3.3.0
djangorestframework>=3.12.4,<3.13.0
Pillow
django-cors-headers